feat: provide a way to extend OSGi specific servlet to support OSGi
fixes #25
