+++ .gitignore
@@ -1,1 +1,4 @@
 **/target/
+**/.project
+**/.classpath
+/driver_zips/
