+++ flow-osgi/.gitignore
@@ -1,2 +1,3 @@
 /target/
 /bin/
+/.classpath
