+++ test-containers/bnd-tools-smoke-test/test-project/.gitignore
@@ -1,0 +1,8 @@
+/node_modules/
+/pnpmfile.js
+/tsconfig.json
+/types.d.ts
+/webpack.config.js
+/webpack.generated.js
+/pnpm-lock.yaml
+/package.json
