+++ test-containers/bnd-tools-smoke-test/test-project/src/main/java/com/vaadin/flow/smoke/osgi/SmokeOSGiVaadinServlet.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright (C) 2020 Vaadin Ltd
+ *
+ * This program is available under Commercial Vaadin Developer License
+ * 4.0 (CVDLv4).
+ *
+ *
+ * For the full License, see <https://vaadin.com/license/cvdl-4.0>.
+ */
+package com.vaadin.flow.smoke.osgi;
+
+import javax.servlet.Servlet;
+import javax.servlet.ServletException;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.http.whiteboard.propertytypes.HttpWhiteboardServletAsyncSupported;
+import org.osgi.service.http.whiteboard.propertytypes.HttpWhiteboardServletPattern;
+
+import com.vaadin.flow.osgi.support.servlet.OSGiVaadinServlet;
+
+@Component(service = Servlet.class)
+@HttpWhiteboardServletAsyncSupported
+@HttpWhiteboardServletPattern("/not-servlet-context-listener-based/*")
+public class SmokeOSGiVaadinServlet extends OSGiVaadinServlet {
+
+    @Override
+    protected void servletInitialized() throws ServletException {
+        getService().setClassLoader(getClass().getClassLoader());
+    }
+}
